+++ IsBasvuru/IsBasvuru.WebAPI/Controllers/RolController.cs
@@ -1,42 +1,37 @@
-﻿using IsBasvuru.Domain.DTOs.AdminDtos.RolDtos;
-using IsBasvuru.Domain.Interfaces;
+﻿using IsBasvuru.Domain.Interfaces;
+using IsBasvuru.WebAPI.Controllers;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
 
-namespace IsBasvuru.WebAPI.Controllers
+[Route("api/[controller]")]
+[ApiController]
+[Authorize(Roles = "SuperAdmin,Admin")]
+public class RolController : BaseController
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class RolController : BaseController // 1. BaseController'dan miras al
+    private readonly IRolService _service;
+    public RolController(IRolService service) { _service = service; }
+
+    [HttpGet("GetAll")]
+    public async Task<IActionResult> GetAll()
     {
-        private readonly IRolService _service;
+        var response = await _service.GetAllAsync();
+        return CreateActionResultInstance(response);
+    }
 
-        public RolController(IRolService service)
-        {
-            _service = service;
-        }
+    [HttpGet("GetById/{id}")]
+    public async Task<IActionResult> GetById([FromRoute] int id)
+    {
+        if (id <= 0) return BadRequest("Geçersiz ID.");
+        var response = await _service.GetByIdAsync(id);
+        return CreateActionResultInstance(response);
+    }
 
-        [HttpGet]
-        public async Task<IActionResult> GetAll()
-        {
-            var response = await _service.GetAllAsync();
-            return CreateActionResultInstance(response);
-        }
-
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetById(int id)
-        {
-            var response = await _service.GetByIdAsync(id);
-            return CreateActionResultInstance(response);
-        }
-
-       
-
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(int id)
-        {
-            var response = await _service.DeleteAsync(id);
-            return CreateActionResultInstance(response);
-        }
+    [HttpDelete("Delete/{id}")]
+    [Authorize(Roles = "SuperAdmin")]
+    public async Task<IActionResult> Delete([FromRoute] int id)
+    {
+        if (id <= 0) return BadRequest("Geçersiz ID.");
+        var response = await _service.DeleteAsync(id);
+        return CreateActionResultInstance(response);
     }
 }
